% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boro_plot.R
\name{boro_plot}
\alias{boro_plot}
\title{Funkcja wyswietla wykres na podstawie wyniku funkcji "fun_pack"}
\usage{
boro_plot(wynik = out)
}
\arguments{
\item{wynik}{\itemize{
\item data frame, jest to wynik funkcji "fun_pack"
}}
}
\value{
plot
}
\description{
Przyjmuje nastepujace argumenty:
}
